🛠️ :: ReviewApi 리팩토링

+++ data/src/main/java/team/retum/data/remote/response/review/FetchReviewsResponse.kt
@@ -10,7 +10,7 @@ data class FetchReviewsResponse(
 
 data class Review(
     @SerializedName("review_id") val reviewId: String,
-    @SerializedName("year") val year: Int,
+    @SerializedName("year") val year: Long,
     @SerializedName("writer") val writer: String,
 )
